+++ client/grpc/tmservice/service_test.go
@@ -152,7 +152,7 @@ func (s IntegrationTestSuite) TestQueryBlockResultsByHeight() {
 	s.Require().Equal(0, len(txResult))
 
 	beginBlock := blockResultsRes.GetResBeginBlock()
-	s.Require().Equal(11, len(beginBlock.Events)) // coinbase event (6) + transfer mintModule to feeCollectorName(5)
+	s.Require().Equal(9, len(beginBlock.Events)) // coinbase event (6) + transfer mintModule to feeCollectorName(3)
 
 	endBlock := blockResultsRes.GetResEndBlock()
 	s.Require().Equal(0, len(endBlock.Events))