+++ x/collection/keeper/send.go
@@ -21,7 +21,6 @@ func (k Keeper) SendCoins(ctx sdk.Context, contractID string, from, to sdk.AccAd
 					From:       from.String(),
 					To:         to.String(),
 				}
-				ctx.EventManager().EmitEvent(collection.NewEventOperationTransferNFT(event))
 				if err := ctx.EventManager().EmitTypedEvent(&event); err != nil {
 					panic(err)
 				}