+++ x/foundation/keeper/internal/keeper_test.go
@@ -133,8 +133,14 @@ func (s *KeeperTestSuite) SetupTest() {
 		s.Require().NoError(err)
 	}
 
-	// create a proposal
-	activeProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{newMsgCreateDog("shiba1")})
+	// create an active proposal, voted yes by all members except the first member
+	activeProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{
+		&foundation.MsgWithdrawFromTreasury{
+			Authority: s.authority.String(),
+			To:        s.stranger.String(),
+			Amount:    sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, s.balance)),
+		},
+	})
 	s.Require().NoError(err)
 	s.activeProposal = *activeProposal
 
@@ -147,8 +153,8 @@ func (s *KeeperTestSuite) SetupTest() {
 		s.Require().NoError(err)
 	}
 
-	// create a proposal voted by all members
-	votedProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{newMsgCreateDog("shiba2")})
+	// create a proposal voted no by all members
+	votedProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{newMsgCreateDog("shiba1")})
 	s.Require().NoError(err)
 	s.votedProposal = *votedProposal
 
@@ -162,7 +168,7 @@ func (s *KeeperTestSuite) SetupTest() {
 	}
 
 	// create an withdrawn proposal
-	withdrawnProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{newMsgCreateDog("shiba3")})
+	withdrawnProposal, err := s.impl.SubmitProposal(s.ctx, []string{s.members[0].String()}, "", []sdk.Msg{newMsgCreateDog("shiba2")})
 	s.Require().NoError(err)
 	s.withdrawnProposal = *withdrawnProposal